Proper way of tokenizing api requests
