+++ package.json
@@ -10,6 +10,7 @@
     "css-loader": "^0.19.0",
     "extract-text-webpack-plugin": "^0.8.2",
     "file-loader": "^0.8.4",
+    "history": "^1.11.1",
     "isomorphic-fetch": "^2.1.1",
     "jade": "^1.11.0",
     "jade-html-loader": "bline/jade-html-loader",
