+++ src/Services/api.js
@@ -1,4 +1,5 @@
 import fetch from 'isomorphic-fetch'
+import {getToken} from './auth'
 
 const API_HOST = 'http://localhost:3000'
 
@@ -6,7 +7,7 @@ export function getWhoAmI(){
   return fetch(`${API_HOST}/v1/users/whoami`,{
     headers:{
       'Accept': 'application/json',
-      'X-Authentication-Token': window.localStorage.getItem('authToken')
+      'X-Authentication-Token': getToken()
     }
   })
 }
@@ -17,7 +18,7 @@ export function getHome({offset}) {
     {
       headers: {
         'Accept': 'application/json',
-        'X-Authentication-Token': window.localStorage.getItem('authToken')
+        'X-Authentication-Token': getToken()
       }
     }
   )
